+++ CHANGELOG.md
@@ -6,6 +6,20 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+### Changed
+
+### Deprecated
+
+### Fixed
+
+### Removed
+
+### Security
+
+## [1.9.0] - Released 2021-08-30
+
+### Added
+
 - Add support for Event conferencing
 - Add support for Account deletion
 
@@ -147,7 +161,8 @@ fetching all via fetchAll method
 
 Initial preview release
 
-[Unreleased]: https://github.com/nylas/nylas-java/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/nylas/nylas-java/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/nylas/nylas-java/releases/tag/v1.9.0
 [1.8.0]: https://github.com/nylas/nylas-java/releases/tag/v1.8.0
 [1.7.0]: https://github.com/nylas/nylas-java/releases/tag/v1.7.0
 [1.6.0]: https://github.com/nylas/nylas-java/releases/tag/v1.6.0